Moving SelectedItemsList to @fluentui/react-internal package (#15430)
