+++ change/@fluentui-react-examples-2020-10-08-11-10-58-feat-react-internal-selected-items-list.json
@@ -1,0 +1,8 @@
+{
+  "type": "minor",
+  "comment": "Removing react-next SelectedItemsList examples.",
+  "packageName": "@fluentui/react-examples",
+  "email": "czearing@outlook.com",
+  "dependentChangeType": "patch",
+  "date": "2020-10-08T18:10:29.688Z"
+}
