+++ change/@fluentui-react-internal-2020-10-08-11-10-58-feat-react-internal-selected-items-list.json
@@ -1,0 +1,8 @@
+{
+  "type": "minor",
+  "comment": "Moving react-next SelectedItemsList to react-internal package.",
+  "packageName": "@fluentui/react-internal",
+  "email": "czearing@outlook.com",
+  "dependentChangeType": "patch",
+  "date": "2020-10-08T18:10:46.148Z"
+}
