+++ change/@fluentui-react-next-2020-10-08-11-10-58-feat-react-internal-selected-items-list.json
@@ -1,0 +1,8 @@
+{
+  "type": "prerelease",
+  "comment": "Moving react-next SelectedItemsList to react-internal package.",
+  "packageName": "@fluentui/react-next",
+  "email": "czearing@outlook.com",
+  "dependentChangeType": "patch",
+  "date": "2020-10-08T18:10:58.658Z"
+}
