--- packages/react-examples/src/react-next/SelectedPeopleList/docs/SelectedPeopleListDonts.md
@@ -1,1 +1,0 @@
-- Use for other items besides selected people.
