--- packages/react-examples/src/react-next/SelectedPeopleList/docs/SelectedPeopleListDos.md
@@ -1,1 +1,0 @@
-- Use them to represent list of selected people.
