+++ packages/react-internal/src/components/SelectedItemsList/BaseSelectedItemsList.test.tsx
@@ -4,6 +4,7 @@ import * as renderer from 'react-test-renderer';
 
 import { IBaseSelectedItemsListProps, ISelectedItemProps } from './BaseSelectedItemsList.types';
 import { BaseSelectedItemsList } from './BaseSelectedItemsList';
+import { isConformant } from '../../common/isConformant';
 
 export interface ISimple {
   key: string;
@@ -26,6 +27,12 @@ describe('SelectedItemsList', () => {
       const component = renderer.create(<BaseSelectedItemsListWithType />);
       const tree = component.toJSON();
       expect(tree).toMatchSnapshot();
+    });
+
+    isConformant({
+      Component: BaseSelectedItemsList,
+      displayName: 'BaseSelectedItemsList',
+      disabledTests: [`has-top-level-file`],
     });
 
     it('can remove items', () => {
