+++ packages/react-internal/src/components/SelectedItemsList/BaseSelectedItemsList.tsx
@@ -3,8 +3,9 @@ import { Selection } from '../../Selection';
 
 import { IBaseSelectedItemsList, IBaseSelectedItemsListProps, ISelectedItemProps } from './BaseSelectedItemsList.types';
 import { initializeComponentRef } from '../../Utilities';
+import { IObjectWithKey } from '../../Utilities';
 
-export interface IBaseSelectedItemsListState<T = any> {
+export interface IBaseSelectedItemsListState<T> {
   items: T[];
 }
 
@@ -12,7 +13,15 @@ export class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>>
   extends React.Component<P, IBaseSelectedItemsListState<T>>
   implements IBaseSelectedItemsList<T> {
   protected root: HTMLElement;
-  protected selection: Selection;
+  private _defaultSelection: Selection;
+
+  public static getDerivedStateFromProps(newProps: IBaseSelectedItemsListProps<any>) {
+    if (newProps.selectedItems) {
+      return { items: newProps.selectedItems };
+    }
+
+    return null;
+  }
 
   constructor(basePickerProps: P) {
     super(basePickerProps);
@@ -24,9 +33,7 @@ export class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>>
     };
 
     // Create a new selection if one is not specified
-    this.selection = this.props.selection
-      ? (this.props.selection as Selection)
-      : new Selection({ onSelectionChanged: this.onSelectionChanged });
+    this._defaultSelection = new Selection({ onSelectionChanged: this.onSelectionChanged });
   }
 
   public get items(): T[] {
@@ -88,8 +95,8 @@ export class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>>
 
   public removeItems = (itemsToRemove: any[]): void => {
     const { items } = this.state;
-    const itemsCanRemove = itemsToRemove.filter((item: any) => this._canRemoveItem(item));
-    const newItems: T[] = items.filter((item: any) => itemsCanRemove.indexOf(item) === -1);
+    const itemsCanRemove = itemsToRemove.filter(item => this._canRemoveItem(item));
+    const newItems: T[] = items.filter(item => itemsCanRemove.indexOf(item) === -1);
     const firstItemToRemove = itemsCanRemove[0];
     const index: number = items.indexOf(firstItemToRemove);
 
@@ -132,6 +139,12 @@ export class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>>
     return this.selection.getSelectedCount() > 0;
   }
 
+  public componentDidUpdate(oldProps: P, oldState: IBaseSelectedItemsListState<IObjectWithKey>): void {
+    if (this.state.items && this.state.items !== oldState.items) {
+      this.selection.setItems(this.state.items);
+    }
+  }
+
   public unselectAll(): void {
     this.selection.setAllSelected(false);
   }
@@ -140,26 +153,12 @@ export class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>>
     return this.selection.getSelection() as T[];
   }
 
-  public UNSAFE_componentWillUpdate(newProps: P, newState: IBaseSelectedItemsListState): void {
-    if (newState.items && newState.items !== this.state.items) {
-      this.selection.setItems(newState.items);
-    }
-  }
-
   public componentDidMount(): void {
     this.selection.setItems(this.state.items);
   }
 
-  public UNSAFE_componentWillReceiveProps(newProps: P): void {
-    const newItems = newProps.selectedItems;
-
-    if (newItems) {
-      this.setState({ items: newItems });
-    }
-
-    if (newProps.selection) {
-      this.selection = newProps.selection;
-    }
+  protected get selection(): Selection {
+    return this.props.selection ?? this._defaultSelection;
   }
 
   public render(): any {
