+++ packages/react-internal/src/components/SelectedItemsList/SelectedPeopleList/Items/ExtendedSelectedItem.scss
@@ -1,4 +1,4 @@
-@import '../../../../common/common';
+@import '~@fluentui/common-styles/dist/sass/common';
 
 .personaContainer {
   @include focus-border(-2px);
