+++ packages/react-internal/src/components/SelectedItemsList/SelectedPeopleList/Items/ExtendedSelectedItem.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
+import { IconButton } from '../../../../Button';
 import { css, getId, initializeComponentRef } from '../../../../Utilities';
 import { Persona, PersonaSize } from '../../../../Persona';
 import { ISelectedPeopleItemProps } from '../SelectedPeopleList';
-import { IconButton } from '../../../../Button';
 import * as stylesImport from './ExtendedSelectedItem.scss';
+
 const styles: any = stylesImport;
 
 export interface IPeoplePickerItemState {
