--- packages/react-next/src/components/SelectedItemsList/SelectedPeopleList/Items/EditingItem.scss
@@ -1,13 +1,0 @@
-.editingInput {
-  border: 0px;
-  outline: none;
-  width: 100%;
-
-  &::-ms-clear {
-    display: none;
-  }
-}
-
-.editingContainer {
-  margin: 4px;
-}
