--- packages/react-next/src/components/SelectedItemsList/SelectedPeopleList/Items/ExtendedSelectedItem.scss
@@ -1,150 +1,0 @@
-@import '~@fluentui/common-styles/dist/sass/common';
-
-.personaContainer {
-  @include focus-border(-2px);
-  border-radius: 15px;
-  display: inline-flex;
-  align-items: center;
-  background: $ms-color-themeLighterAlt;
-  margin: 4px;
-  cursor: default;
-  user-select: none;
-  vertical-align: middle;
-  position: relative;
-
-  // Setting global values here to override persona's normal behavior specifically in this selected variation
-  :global(.ms-Persona-primaryText) {
-    color: $ms-color-themeDark;
-    @include ms-font-m;
-
-    &.hover {
-      color: $ms-color-themeDark;
-    }
-
-    @include high-contrast {
-      color: HighlightText;
-    }
-  }
-
-  .actionButton {
-    &:hover {
-      background: $ms-color-themeLight;
-    }
-
-    :global(.ms-Button-icon) {
-      color: $ms-color-themeDark;
-
-      @include high-contrast {
-        color: HighlightText;
-      }
-    }
-  }
-
-  &:hover {
-    background: $ms-color-themeLighter;
-
-    // Setting global values here to override persona's normal behavior specifically in this selected variation
-    :global(.ms-Persona-primaryText) {
-      color: $ms-color-themeDark;
-      @include ms-font-m;
-
-      @include high-contrast {
-        color: HighlightText;
-      }
-    }
-  }
-
-  &.personaContainerIsSelected {
-    background: $ms-color-themePrimary;
-
-    // Setting global values here to override persona's normal behavior specifically in this selected variation
-    :global(.ms-Persona-primaryText) {
-      color: $ms-color-white;
-
-      @include high-contrast {
-        color: HighlightText;
-      }
-    }
-
-    .actionButton {
-      color: $ms-color-white;
-
-      :global(.ms-Button-icon) {
-        color: $ms-color-themeDark;
-
-        &:hover {
-          background: $ms-color-themeDark;
-        }
-
-        @include high-contrast {
-          color: HighlightText;
-        }
-      }
-    }
-
-    @include high-contrast {
-      border-color: Highlight;
-      background: Highlight;
-      -ms-high-contrast-adjust: none;
-    }
-  }
-
-  &.validationError {
-    // Setting global values here to override persona's normal behavior in the error version
-    :global(.ms-Persona-primaryText) {
-      color: $ms-color-red;
-    }
-
-    :global(.ms-Persona-initials) {
-      font-size: 20px;
-    }
-  }
-
-  @include high-contrast {
-    border: 1px solid WindowText;
-  }
-
-  .itemContent {
-    flex: 0 1 auto;
-    min-width: 0px;
-
-    /** CSS below is needed for IE 11 to properly truncate long persona names in the picker **/
-    max-width: 100%;
-  }
-
-  .removeButton {
-    border-radius: 15px;
-    flex: 0 0 auto;
-    width: 33px;
-    height: 33px;
-    flex-basis: 32px;
-  }
-
-  .expandButton {
-    border-radius: 15px 0px 0px 15px;
-    height: 33px;
-    width: 44px;
-    padding-right: 16px;
-    position: inherit;
-    display: flex;
-    margin-right: -17px;
-  }
-
-  .personaWrapper {
-    position: relative;
-    display: inherit;
-
-    :global(.ms-Persona-details) {
-      padding: 0px 8px;
-    }
-  }
-
-  .personaDetails {
-    flex: 0 1 auto;
-  }
-}
-
-.itemContainer {
-  display: inline-block;
-  vertical-align: top;
-}
