--- packages/react-next/src/components/SelectedItemsList/index.ts
@@ -1,5 +1,0 @@
-export * from './BaseSelectedItemsList.types';
-export * from './BaseSelectedItemsList';
-export * from './SelectedPeopleList/SelectedPeopleList';
-export * from './SelectedPeopleList/Items/ExtendedSelectedItem';
-export * from './SelectedPeopleList/Items/EditingItem.types';
